Revert "spectra"
This reverts commit 2d03d88c16d16ad6831c19a7921b84600daeb284.
