--- Content.Shared/Stealth/SharedStealthSystem.Move.cs
@@ -1,30 +1,0 @@
-using Content.Shared.Stealth.Components;
-
-namespace Content.Shared.Stealth;
-
-public abstract partial class SharedStealthSystem
-{
-    private void InitializeMove()
-    {
-        SubscribeLocalEvent<StealthOnMoveComponent, MoveEvent>(OnMove);
-        SubscribeLocalEvent<StealthOnMoveComponent, GetVisibilityModifiersEvent>(OnGetMoveVisibilityModifiers);
-    }
-
-    private void OnMove(EntityUid uid, StealthOnMoveComponent component, ref MoveEvent args)
-    {
-        if (_timing.ApplyingState)
-            return;
-
-        if (args.NewPosition.EntityId != args.OldPosition.EntityId)
-            return;
-
-        var delta = component.MovementVisibilityRate * (args.NewPosition.Position - args.OldPosition.Position).Length();
-        ModifyVisibility(uid, delta);
-    }
-
-    private void OnGetMoveVisibilityModifiers(EntityUid uid, StealthOnMoveComponent component, GetVisibilityModifiersEvent args)
-    {
-        var mod = args.SecondsSinceUpdate * component.PassiveVisibilityRate;
-        args.FlatModifier += mod;
-    }
-}
